pyproject.toml
[tool.poetry]
name = "koopman-control"
version = "0.1.0"
description = ""
authors = ["Jacob Evarts <31416036+jacob-evarts@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
torch = "^2.9.1"
hydra-core = "^1.3.2"
optuna = "^4.6.0"
pytorch-lightning = "^2.6.0"
plotly = "^6.5.2"
optuna-integration = "^4.6.0"
torchvision = "^0.24.1"
wandb = "^0.24.0"
lightning = "^2.6.0"
scikit-learn = "^1.8.0"
kaleido = "^1.2.0"
matplotlib = "^3.10.8"
h5py = "^3.15.1"
tox = "^4.34.1"
black = "^26.1.0"
mypy = "^1.19.1"
pylint = "^4.0.4"
pandas = "^3.0.0"
imageio = "^2.37.2"
torch-geometric = "^2.7.0"


[tool.poetry.group.test.dependencies]
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
